+++ code/src/cljs/sixsq/nuvla/ui/data/events.cljs
@@ -125,8 +125,8 @@
                 ::spec/data-sets
                 ::spec/selected-data-set-ids] :as db} :db} _]
     (let [selected-data-sets (vals (filter (fn [[k v]] (boolean (selected-data-set-ids k))) data-sets))
-          query-application (apply general-utils/join-and (map :module-filter selected-data-sets))
-          query-objects (apply general-utils/join-or (map :data-record-filter selected-data-sets))]
+          query-application  (apply general-utils/join-and (map :module-filter selected-data-sets))
+          query-objects      (apply general-utils/join-or (map :data-record-filter selected-data-sets))]
       {:db                  (assoc db ::spec/application-select-visible? true
                                       ::spec/loading-applications? true
                                       ::spec/selected-application-id nil